requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
tensorflow==2.16.2
numpy==1.26.4
pillow==10.3.0
python-multipart==0.0.6